+++ src/atlas/array/native/NativeMakeView.cc
@@ -51,7 +51,7 @@ template <typename Value, int Rank>
 ArrayView<Value, Rank> make_device_view(Array& array) {
 #if ATLAS_HAVE_GPU
     ATLAS_ASSERT(array.deviceAllocated(),"make_device_view: Array not allocated on device");
-    return ArrayView<Value, Rank>((array.device_data<Value>()), array.shape(), array.strides());
+    return ArrayView<Value, Rank>((array.device_data<Value>()), array.shape(), array.device_strides());
 #else
     return make_host_view<Value, Rank>(array);
 #endif
@@ -61,7 +61,7 @@ template <typename Value, int Rank>
 ArrayView<const Value, Rank> make_device_view(const Array& array) {
 #if ATLAS_HAVE_GPU
     ATLAS_ASSERT(array.deviceAllocated(),"make_device_view: Array not allocated on device");
-    return ArrayView<const Value, Rank>(array.device_data<const Value>(), array.shape(), array.strides());
+    return ArrayView<const Value, Rank>(array.device_data<const Value>(), array.shape(), array.device_strides());
 #else
     return make_host_view<const Value, Rank>(array);
 #endif